pyproject.toml
[tool.poetry]
name = "tomocube_downloader"
version = "0.1.0"
description = ""
authors = ["mdsung <mdskylover@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.49.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
python-dotenv = "^0.20.0"
PyMySQL = "^1.0.2"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
